rm condvar causing deadlock. Still a race cond.
In order to avoid deadlock, it was required that there were n threads
for processing. For example, if n=5 and my four nbrs sent their START
msgs while I still hadn't started, there would be four threads waiting
on the condvar, which for #threads<5 meant there were no resources to
start and signal the condvar.

However, even after this fix there is still a race condition. Auto
assignment seems stable, but when called just right, a new formation
will cause the auctions to become out of sync and CBAA will fail to
converge.

It is unclear, but this may be the cause of assignments where a single
vehicle has bid the highest for two spots, resulting in a segfault.