keep bids in queue until auction is open
For a fully-connected graph, this eliminates the loss of bids by keeping
them in the queue until the auction is open. As a result, (for fc) we do
not see any "Auctioneer is busy!" errors from the autoauctionCb.

However, the number of invalid assignments increases significantly
(especially with auction dt at 0.25).